refactor(infra): move fee decoder logic
